+++ desktop/src/main/java/bisq/desktop/main/portfolio/pendingtrades/steps/buyer/BuyerStep1View.java
@@ -24,6 +24,9 @@ import bisq.desktop.main.portfolio.pendingtrades.steps.TradeStepView;
 import bisq.core.locale.Res;
 import bisq.core.support.dispute.DisputeValidation;
 import bisq.core.trade.bisq_v1.TradeDataValidation;
+import bisq.core.trade.validation.DelayedPayoutTxValidation;
+import bisq.core.trade.validation.exceptions.MissingTxException;
+import bisq.core.trade.validation.exceptions.ValidationException;
 
 public class BuyerStep1View extends TradeStepView {
 
@@ -83,13 +86,13 @@ public class BuyerStep1View extends TradeStepView {
 
     private void validatePayoutTx() {
         try {
-            TradeDataValidation.validateDelayedPayoutTx(trade.getDelayedPayoutTx(),
+            DelayedPayoutTxValidation.validateDelayedPayoutTx(trade.getDelayedPayoutTx(),
                     trade,
                     model.dataModel.btcWalletService);
-        } catch (TradeDataValidation.MissingTxException ignore) {
+        } catch (MissingTxException ignore) {
             // We don't react on those errors as a failed trade might get listed initially but getting removed from the
             // trade manager after initPendingTrades which happens after activate might be called.
-        } catch (TradeDataValidation.ValidationException | DisputeValidation.ValidationException e) {
+        } catch (ValidationException | DisputeValidation.ValidationException e) {
             if (!model.dataModel.tradeManager.isAllowFaultyDelayedTxs()) {
                 new Popup().warning(Res.get("portfolio.pending.invalidTx", e.getMessage())).show();
             }
@@ -100,7 +103,7 @@ public class BuyerStep1View extends TradeStepView {
     private void validateDepositInputs() {
         try {
             TradeDataValidation.validateDepositInputs(trade);
-        } catch (TradeDataValidation.ValidationException e) {
+        } catch (ValidationException e) {
             if (!model.dataModel.tradeManager.isAllowFaultyDelayedTxs()) {
                 new Popup().warning(Res.get("portfolio.pending.invalidTx", e.getMessage())).show();
             }
