+++ desktop/src/main/java/bisq/desktop/main/portfolio/pendingtrades/steps/buyer/BuyerStep2View.java
@@ -110,8 +110,10 @@ import bisq.core.payment.payload.SwiftAccountPayload;
 import bisq.core.payment.payload.USPostalMoneyOrderAccountPayload;
 import bisq.core.payment.payload.WesternUnionAccountPayload;
 import bisq.core.support.dispute.DisputeValidation;
-import bisq.core.trade.bisq_v1.TradeDataValidation;
 import bisq.core.trade.model.bisq_v1.Trade;
+import bisq.core.trade.validation.DelayedPayoutTxValidation;
+import bisq.core.trade.validation.exceptions.MissingTxException;
+import bisq.core.trade.validation.exceptions.ValidationException;
 import bisq.core.user.DontShowAgainLookup;
 import bisq.core.util.VolumeUtil;
 
@@ -738,13 +740,13 @@ public class BuyerStep2View extends TradeStepView {
 
     private void validatePayoutTx() {
         try {
-            TradeDataValidation.validateDelayedPayoutTx(trade.getDelayedPayoutTx(),
+            DelayedPayoutTxValidation.validateDelayedPayoutTx(trade.getDelayedPayoutTx(),
                     trade,
                     model.dataModel.btcWalletService);
-        } catch (TradeDataValidation.MissingTxException ignore) {
+        } catch (MissingTxException ignore) {
             // We don't react on those errors as a failed trade might get listed initially but getting removed from the
             // trade manager after initPendingTrades which happens after activate might be called.
-        } catch (TradeDataValidation.ValidationException | DisputeValidation.ValidationException e) {
+        } catch (ValidationException | DisputeValidation.ValidationException e) {
             if (!model.dataModel.tradeManager.isAllowFaultyDelayedTxs()) {
                 new Popup().warning(Res.get("portfolio.pending.invalidTx", e.getMessage())).show();
             }
